remove ts it was causing far too many issues to the project
