+++ package.json
@@ -5,24 +5,20 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "dev": "node --loader ts-node/esm index.ts",
-    "start": "ts-node index.js",
+    "start": "NODE_ENV=development nodemon index.js",
     "build": "tsc -p .",
     "style": "stylus -w  ./public/main.styl  -o ./public"
   },
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@types/express": "^4.17.13",
-    "@types/node": "^18.0.0",
     "dotenv": "^16.0.1",
-    "ts-node": "^10.8.1",
-    "typescript": "^4.7.4",
     "wifi-qr-code-generator": "^1.1.1"
   },
   "dependencies": {
     "express": "^4.18.1",
-    "jq": "^1.7.2",
+    "morgan": "^1.10.0",
+    "nodemon": "^2.0.18",
     "pug": "^3.0.2",
     "stylus": "^0.58.1"
   }
