+++ public/main.css
@@ -24,7 +24,7 @@ body {
 body {
   margin: 0;
   background-color: #efefef;
-  background-image: url("./img/qr1.jpg");
+  background-image: url("./img/qr.jpg");
   background-repeat-x: no-repeat;
   background-size: cover;
   padding: 5rem 0;
@@ -165,28 +165,8 @@ li {
 .c-w-text {
   color: #fff;
 }
-.sidebar {
-  min-width: 15rem;
-  visibility: hidden;
-}
-.sidebar label {
-  color: #fff;
-}
 .m-w-50 {
   max-width: 50%;
-}
-button.primary {
-  background-color: #1a1a1a;
-  color: #fff;
-  border: none;
-}
-.panel {
-  background-color: #1a1a1a;
-  color: #fff;
-  margin-bottom: 1rem;
-  padding: 1rem;
-  opacity: 0.9;
-  border-radius: 0.25rem;
 }
 .max-width-23 {
   max-width: 23rem;
@@ -226,18 +206,6 @@ button.primary {
 }
 .material-icons.small {
   font-size: 17px;
-}
-.background {
-  display: block;
-  position: absolute;
-  top: 0;
-  left: 0;
-  -o-object-fit: cover;
-  object-fit: cover;
-  width: 100%;
-  height: 100%;
-  -webkit-mask-image: radial-gradient(#fff 0%, #fff 30%, transparent 80%, transparent);
-  mask-image: radial-gradient(#fff 0%, #fff 30%, transparent 80%, transparent);
 }
 main {
   max-width: 30%;
