+++ public/main.styl
@@ -26,7 +26,7 @@ html, body
 body
     margin: 0;
     background-color: #efefef;
-    background-image: url(./img/qr1.jpg);
+    background-image: url(./img/qr.jpg);
     background-repeat-x: no-repeat;
     background-size: cover;
     padding: 5rem 0;
@@ -56,24 +56,6 @@ input
         border-color: rgb(74, 74, 199)
     &.error
         border-color: #ff0000
-
-
-// button
-//     padding: .5rem
-//     cursor: pointer
-//     appearance: none
-//     outline: none
-//     border-radius: 3px
-//     background-color: #e7d2cc
-//     border: 2px solid transparent
-//     color: #2e2e2e
-//     font-weight: 600
-//     transition: all .2s ease-in-out
-
-// button:hover
-//     border: 2px solid #e7d2cc
-//     background-color: #fff
-
 
 input[type="radio"]
     cursor: pointer
@@ -179,32 +161,8 @@ ul,li
 .c-w-text
     color: c-w
 
-.sidebar
-    min-width: 15rem
-    visibility: hidden
-
-.sidebar label
-    color: c-w
-
 .m-w-50
     max-width: 50%
-
-button.primary
-    background-color: primary;
-    color: #ffffff
-    border: none
-
-// .radio-list
-//     margin-left: 10rem
-//     margin-right: 17rem
-
-.panel
-    background-color: primary
-    color: c-w
-    margin-bottom: 1rem;
-    padding: 1rem
-    opacity: panel-opacity
-    border-radius: .25rem
 
 .max-width-23
     max-width: 23rem
@@ -250,20 +208,7 @@ button.primary
 .material-icons.small
     font-size: 17px
 
-//
-.background
-    display: block
-    position: absolute
-    top: 0
-    left: 0
-    -o-object-fit: cover
-    object-fit: cover
-    width: 100%
-    height: 100%
-    -webkit-mask-image: radial-gradient(white 0%, white 30%, transparent 80%, transparent)
-    mask-image: radial-gradient(white 0%, white 30%, transparent 80%, transparent)
-
-    //New
+// Qr code styles
 
 main
     max-width: 30%;
@@ -317,9 +262,6 @@ small
         cursor: pointer
 
 // Gradients
-
-// remCalc(val)
-// return (val/16)rem
 
 gradients = {
 	one 	: linear-gradient(135deg, #ecc344 0%, #e42d7f 100%), 		// yellow pink
